requirements.txt
fastapi
uvicorn
transformers
torch
nltk
sentence-transformers
python-docx
pypdf2
python-multipart